+++ CHANGELOG.md
@@ -10,6 +10,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 > *Note*: Some version numbers may be skipped to align feature set with the Node.js version.
 
 ---
+## Version 4.2.32, 4/24/2025
+
+### Added
+
+N/A
+
+### Removed
+
+N/A
+
+### Changed
+
+1. Improvement in custom content-type resolver
+2. OSS version update
+
+```
+vertx 4.1.14
+guava 33.4.8-jre 
+junit5-bom 5.12.2
+kotlin 2.1.20
+Kafka client 4.0.0
+gson 2.13.0
+```
+---
 ## Version 4.2.31, 4/23/2025
 
 ### Added